+++ helper-scripts/deleteStacksZx/functions/delete_stack.sh
