+++ helper-scripts/deleteStacksZx/functions/formatting.ts
