+++ helper-scripts/deleteStacksZx/functions/getStacksToDelete.ts
