+++ helper-scripts/deleteStacksZx/functions/mockScriptUtils.ts
