+++ helper-scripts/deleteStacksZx/functions/protectedStacks.ts
