+++ helper-scripts/deleteStacksZx/functions/validateStacks.ts
